Script Update, not Reproducible for other PCs.

+++ changetheme.sh
@@ -1,2 +1,17 @@
-killall waybar
-killall hybridbar
+#! /bin/bash
+if [ $# -eq 1 ]
+then
+	cd ~/sway-dotfiles-script/configs
+	if [ -d "$1" ]
+	then
+		cd $1
+		cp -r config/* ~/.config
+		killall hybridbar
+		killall waybar
+		swaymsg reload
+	fi
+else
+	echo "No arguments given."
+
+fi
+
